+++ lib/components/home/TerminalWrapper.tsx
