+++ lib/utils/api.ts
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import config from '../../../config.json';
+import config from '../../config.json';
 
 export const getProjects = async () => {
   const { data } = await axios.get(
