+++ lib/utils/bin/commands.ts
@@ -1,7 +1,7 @@
 // List of commands that do not require API calls
 
 import * as bin from './index';
-import config from '../../../../config.json';
+import config from '../../../config.json';
 
 // Help
 export const help = async (args: string[]): Promise<string> => {
