+++ lib/utils/bin/index.ts
