+++ lib/utils/bin/sumfetch.ts
@@ -1,4 +1,4 @@
-import config from '../../../../config.json';
+import config from '../../../config.json';
 
 const sumfetch = async (args: string[]): Promise<string> => {
   if (config.ascii === 'zornwebdev') {
