+++ lib/utils/commandExists.ts
