+++ lib/utils/shell.ts
