+++ lib/utils/tabCompletion.ts
